Update patch set 2
Patch Set 2:

No plug-ins that I know of. Except perhaps the changes in the Gerrit-RabbitMQ-Plugin I'm working on. I guess this will end up in a major version upgrade, so I think it should be OK anyways. In the end, CurrentUser is a base class to IdentifiedUser, so if anyone uses it, it should should hopefully still work(?)

Patch-set: 2