Update patch set 2
Patch Set 2:

> > No plug-ins that I know of. Except perhaps the changes in the
 > Gerrit-RabbitMQ-Plugin I'm working on
 > 
 > Are working on rabbitMQ-plugin so it does not by-pass change
 > visibility?

Exactly.

My first plan was to support authGroup (for consistency with the replication plug-in). But due to the use of IdentifiedUser, this proved a bit problematic. So until this change is in, I have another approach; If you specify the following in rabbitmq.config:

  [security]
    authUser = jenkins_verifier

It will filter out all events that is not visible for that Gerrit-user. I figured, this way it will be easy to coordinate events with various Jenkins-clusters. E.g. events for changes that the jenkins_verifier cannot fetch will not even be sent to that rabbitmq queue. This would minimize the amount of triggered bad builds due to Access Rights mismatches in Gerrit.

I'm almost done with this feature. I can give you a ping when done.

Thanks for verifying!

Patch-set: 2
Reviewer: Gerrit User 1006454 <1006454@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0