Update patch set 4
Patch Set 4:

Rebased on top of the change with the test case, and updated to refer to custom built artifact from the Gerrit storage bucket.

I've run the server tests with and without the updated parboiled-java and verified that before the update it fails, and after the update it passes.

You might want to squash these two changes together.

Patch-set: 4