Update patch set 1
Patch Set 1: Code-Review+2

It's already kind of a pain to manually order labels in project.config when inheritance is involved. IIRC inherited labels come before child labels, so if you want to take an inherited label and reorder it, you have to copy the entire block into your project.config.

If we really cared we would probably have an explicit sort order or listing in project.config, and that would be mirrored verbatim in either the project config or the change info as you indicated.

Patch-set: 1
Label: Code-Review=+2