Update patch set 1
Patch Set 1:

> > I just remembered that this had already been done by David but
 > was
 > > abandoned.
 > >
 > > https://gerrit-review.googlesource.com/#/c/61114/
 > 
 > It was abandoned because the idea of removing Edit Message button
 > was rejected by Dave Borowitz and Edwin. So I am wondering why this
 > change was merged without their re-votes? For the record, i agree,
 > to get rid of this button.

Yes, I remember not liking the removal of that button because lots of users got used to it. On the other side I see the value of reducing the complexity of the change screen by offering less functionality there. Inline editing is really good and maybe it's a good way to introduce it to users if they are forced to use it for editing commit messages also. So I think I'm now fine with this change :-)

Patch-set: 1
Reviewer: Gerrit User 1003935 <1003935@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0