Update patch set 1
Patch Set 1:

(3 comments)

> I was going to do a similar change, but change the
 > DeleteDraft{Change|PatchSet} messages to be "cannot delete draft".

Please do, if you have a strong opinion about this. For me this message is not that important.

Still I think it's a difference if draft operations fail due to project access rights and global capabilities or due to global server configuration, and in my opinion this should be reflected in the error message.

Patch-set: 1
Label: Verified=0