Update patch set 15
Patch Set 15:

> 
 > This change seems to be independent of the "submit whole topic"
 > functionality.  Rebase it out of the stack so it can be submitted
 > separately?

Yes it looks quite different. Dave and me had a discussion in the office (sorry for being off list) about the future of the `submitWholeTopic` topic. 

We need to know when to enable the submit button. Therefore we need to know the state of each other change in the same topic. So let's ask the database for each change. (That may take significant additional time if you have a topic of more than just 3-5 changes?). 

So this introduces additional latency and here we already have users complaining about latency. The idea is to separate the loading of the revision action buttons from loading of the change screen. So the change screen can load fast depending only on the database query of this one change, only the revision action button may show up later depending on the speed of the database querying for all the changes in the topic.

This is why I added that to this topic.

Patch-set: 15
Label: Verified=0