Update patch set 6
Patch Set 6:

> Acceptance test compilation fails.
 > 
 > gerrit-acceptance-tests/src/test/java/com/google/gerrit/acceptance/rest/change/AbstractSubmit.java:107:
 > error: cannot find symbol
 > hooks.addEventListener(new EventListener() {
 > ^
 > symbol:   method addEventListener(<anonymous com.google.gerrit.common.EventListener>,com.google.gerrit.server.CurrentUser)
 > location: variable hooks of type com.google.gerrit.common.ChangeHooks
 > Errors: 1. Warnings: 0.
 > BUILD FAILED: //gerrit-acceptance-tests/src/test/java/com/google/gerrit/acceptance/rest/change:submit_util
 > failed with exit code 1:
 > javac

Thanks.  I got a fix for this, so don't work on it.  This reflects a problem that I am going to then face with the next change.  I am not likely to get the next change in this series to work for the tests until the EventDispatcher interface is defined, but that doesn't come until later in the series!  So I think I will need to reorder this series abit from this change down, potentially including this change.  The two serialization changes before this change should be fine, they mostly stand alone.  I will need to move the EventDispatcher interface definition before I split this class in 2.

Patch-set: 6
Label: Verified=0