Update patch set 1
Patch Set 1:

(1 comment)

> (1 comment)
 > 
 > I can see one possible problem. If someone asks for a rebased patch
 > for whatever reason (maybe even "I prefer linear history" thing),
 > then the reviewer might be left waiting even after a patch is
 > correctly rebased. I cannot judge how much of a problem this is.
 > 
 > That said, I like this feature in general.

Actually, I think no configuration option and hardcoded as is is the best.

The biggest problem with Gerrit in regards to email is that reviewers are inundated. 

It's better to err on the side of sending out too few emails. The owner can manually knudge reviewers with an email if necessary.

IMO, a 'knudge' feature doesn't make sense, because it just looks like more Gerrit noise.

Also, adding configuration options here is wrong IMO because it would clutter the Gerrit manual with extremely obscure cases.

Patch-set: 1
Label: Verified=0