Update patch set 23
Patch Set 23: -Code-Review -Verified

Actually, I'm not sure if this is working as expected.

When submitWholeTopic is enabled, and I call this on a change that is part of a topic, it returns the submit action as:

  "submit": {
    "method": "POST",
    "label": "Submit",
    "title": "Submit patch set 1 into master",
    "enabled": true
  },

where I expected it to return "Submit all X changes of the same topic".

Another glitch I noticed is that on the UI the tooltip is showing "Submit all 4 changes of the same topic", but there are actually only 3 open changes on the topic.

Patch-set: 23
Label: Code-Review=0