Update patch set 1
Patch Set 1:

With this I've found a possible problem with the EventTypes: If a plugin registers an event type, the registration is still there even after the plugin is removed or disabled.

To reproduce this, run Gerrit with the replication plugin, run this command, remove the replication plugin, wait till you see the unload message in the console log, and then run the command again. Then you see that the replication plugin's events are still there.

Patch-set: 1