Update patch set 16
Patch Set 16:

I am not quite familiar with the gerrit (it's a huge project when working on it for only 2 months already), so abandon and follow up are change actions and rebase and cherrypick are revision actions, right?

Maybe the callback for revision actions displaying silently fails because it is faster than the clean initialisation of the revision action buttons?

Patch-set: 16