Update patch set 2
Patch Set 2:

(4 comments)

@David: thanks for stepping into the game :-) I really like the idea *BUT* this is too similar to my earlier implementation, which reuses the HTTP login logic.

I was under the impression that Shawn did not like it because of the complexity of the solution *and* the resulting configuration mess.

Shall we try to simplify the approach avoiding the reuse of HTTP Login mechanism? Shall we double-check with Shawn before refining the solution?

Patch-set: 2
Label: Verified=0