Update patch set 1
Patch Set 1:

This is not needed for any of the libs that we currently use.

I implemented this as a workaround when I was trying to use antlr4 version 4.3, because some of the needed classes are only in the jar file with `-complete` suffix. This is not the case with antrl4 version 4.5 though.

Patch-set: 1
Label: Verified=0