Update patch set 3
Patch Set 3:

We try to sync this with maven_jar bucklet [1]. It would be nice to apply classifier change on it as well. In case we would use it in core and a plugin that supports standalone build mode would depend on classifier artifact.

* [1] https://gerrit.googlesource.com/bucklets/+/master/maven_jar.bucklet

Patch-set: 3
Reviewer: Gerrit User 1011323 <1011323@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0