Update patch set 2
Patch Set 2:

> > Is this really critical for 2.10?
 > 
 > Yes. See Shawn's announcement here why: [1].
 > 
 > [1] https://groups.google.com/forum/#!topic/repo-discuss/Hjn-6BV3KBU

I know about this announcement.

The reason I am asking is that on 03.03.2015 you asked about 2.10.1 and mentioned that OpenID related changes were already in the stable-2.10.
Therefore, I understood you that what was in stable-2.10 on 03.03.2015 was enough to release 2.10.1 (regarding the OpenID issue).

This changes was uploaded a day later: 04.03.2015. Therefore, my original question.

Patch-set: 2