Update patch set 10
Patch Set 10:

(2 comments)

Thinking about this further, I do not think this is the right approach for the ChangeMergeQueue to have on locking.

Before we were locking per branch and if the lock was already taken, we'd just need to be sure to reschedule just to be sure.

This approach tries to lock all affected branches at once and if it doesn't succeed we're not sure what to do exactly.

Patch-set: 10