Update patch set 1
Patch Set 1:

> Does this problem occurs on master as well?

yes, exact same error as reported in the bug. Every time we add bindings to Daemon, we must add them to WebAppInitializer but it was not done for the SecureStore.

I had a quick look at how the bindings are done for the SecureStore and it is not that simple to understand what need to be added in the WebAppInitializer to get gerrit to start.

Patch-set: 1