Update patch set 3
Patch Set 3: Code-Review-1

I like Martin's suggestion in principal.  As David pointed there have been a few attempts to add the WIP workflow to Gerrit core however it was denied each time.  That's why it's done in a plugin on this attempt.  Providing a general API for modifying search strings seems like a good idea however I don't see many use cases for that functionality and it seems like it would be a big change.

The thing I don't like about this is change is that it allows a 'wip' query by simply disabling drafts.  Setting 'allowDrafts=false' means disable drafts, it doesn't mean enable 'wip' so why should there be a 'wip' query?   Also hard coding the 'wip' query string would mean that we would need to document that specific search string in Gerrit core which doesn't make sense. 

The WIP plugin is very specific in that it overrides the draft workflow so maybe a compromise for what Martin is asking for is to allow users to override the draft query string in gerrit.config (maybe 'draftQueryString=wip') and use that variable instead of hard coding 'wip' for the query?    Then the WIP plugin's documentation can say something like.. "add 'draftQueryString=wip' to allow querying for wip changes with 'status:wip' or 'is:wip'.

Patch-set: 3
Label: Code-Review=-1