Update patch set 5
Patch Set 5:

That merge was a bit fast, as the edit of Jonathan is not incorporated. :(

The patch set 5 is a based on patch set 3 (ommitting the changes of 4) and just including the commit on ps 1.

Patch-set: 5
Reviewer: Gerrit User 1017837 <1017837@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0