Update patch set 1
Patch Set 1:

> This doesn't take into account if the commit is a merge commit.
 > I don't know if and how this should be handled.
 > I'll look into this further tomorrow.

Not sure what you mean here... The behavior is identical to stable-2.11 for merge commits, meaning that there is no Rebase button and the orange ball next to Parent(s) is not rendered.

Nevertheless, it might be a good idea to start indicating the possibility to rebase (manually) merge commits as well to the users. However, I think it should be done in a separate change to simplify potential back-ports to 2.11.

Patch-set: 1