Update patch set 8
Patch Set 8: Code-Review-1

The dep graph for the UI is much bigger than it should be.

For example JavaEWAH is being reported as a dep of the UI. JavaEWAH is not compiled into the UI. Its not a dep. There must be an edge somewhere in the build graph that is incorrect.

Patch-set: 8
Label: Code-Review=-1