Update patch set 1
Patch Set 1:

> > I would not like to always show the plugin name because the user
 > > doesn't care from which plugin the functionality comes,
 > 
 > Really?

Yes, I think so.

Menu entries that plugins contribute to the top menu do also not contain the plugin name. E.g. the serviceuser plugin adds

* 'People' > 'Create Service User' and
* 'People' > 'List Service Users'

It would not add any value to show the plugin name in addition:

* 'People' > 'Create Service User (serviceuser)' and
* 'People' > 'List Service Users (serviceuser)'

Userd just want to create service users, they don't care if this functionality was implemented by a plugin or by Gerrit core. This is an implementation detail. On the opposite the longer menu item names are confusing and consume more space.

So if we handle it for top menu like this, why should it be different for the user settings menu?


 > What if the plugin author named her plugin's custom
 > setting: "Setting"?

Make a contribution to this plugin to fix the menu name ;-)
We cannot really protect against plugin authors doing bad things.

 > And 17 plugins are installed on the site, how
 > is that clear what the content of that "Setting" screen is trying
 > to achieve?

I already agreed that in case of a name collision we should show the plugin name in addition, so this will be:

* Settings (plugin1)
* Settings (plugin2)
...
* Settings (plugin17)

I will push a fix for this in a minute.

 > Also note, that everything what plugins currently
 > expose always prefixed by plugin name, like capabilities, links,
 > etc. That was the reason, to make plugin name stable, by including
 > it into MANIFEST, so that it is not going to change when JAR file
 > name was changed.

Well not everything is prefixed with the plugin name (e.g. menu entries in the top menu, window title of plugin screens etc.).

Patch-set: 1