Update patch set 12
Patch Set 12:

(1 comment)

Thanks for following up on this.  It looks like the change needs to be rebased on the latest master, though.

Also it might be better if you add your additional changes/fixes in a separate commit so that you get attribution in the git log.  As it is now, it is all attributed to the original author.

Patch-set: 12
Label: Verified=0