Update patch set 12
Patch Set 12:

I can't really separate the changes easily. I did already rebase this, but it's probably out of date yet again. This set includes both fixes to the original and new features. I could just drop the original author and include it as a tag below in the tag section? Originally I didn't change the author because it was only a rebase, but it may be different question now as I have heavily re-worked from the original commit.

Patch-set: 12