Update patch set 1
Patch Set 1:

> I would have thought this would happen entirely client-side.

That was my first thought as well. But then I remembered how I broke the UI left and right when adding SubmitWholeTopic (and the necessarry refactoring before and after), not paying attention to the small details as I should have. 

Doing this on the server side allows to write/adapt tests with ease. (The UI code is only tested manually, while the REST API calls can be exercised easily in acceptance tests)

Patch-set: 1
Label: Verified=0