Update patch set 11
Patch Set 11:

> I have motivation problem to address your comments, though.

I understand frustration if a review of a change takes that long as for this change. I know this change is around for a long time and I only now found time to look at it...

 > I'm sure, no matter what I'm doing here (and on other changes that
 > were once down voted by Shawn and are ready now to be approved for
 > months, like: [1], [2]), you (and other maintainers for that matter)
 > are not going to approve these changes.

It is true, if a maintainer gave a negative vote, I prefer to get a confirmation from that maintainer that his comments were addressed. At least in this case I'm not submitting changes immediately but try to give this maintainer time to have another look. By the way the same is true for any negative vote of a non-maintainer. If this feedback doesn't come in a reasonable time, it depends on the change and how comfortable I am with the touched code area whether I can approve it or not. It also depends on my own workload and time I have for reviews.

 > Shawn is busy with other duties and apparently cannot allocate any
 > time for reviewing pending changes, so why bother?

I don't think that we totally depend on Shawn to get changes in. Actually for this change I see good chances to get it submitted soon.

 > * [1] https://gerrit-review.googlesource.com/67860
 > * [2] https://gerrit-review.googlesource.com/61601

These changes depend on the 'inline-4' topic. I would have already submitted the first changes of the series if they would not have been bundled as a topic which must be submitted together. The topic is stuck because some changes need the Library-Compliance vote [1,2]. To move forward with them I have just pinged those changes the day before yesterday. I cannot vote on this label, but I think we could get the first changes in if we remove the topic and sort the changes so that the ones which require Library-Compliance are on top so that they don't block the others.

[1] https://gerrit-review.googlesource.com/67904
[2] https://gerrit-review.googlesource.com/67905

 > The maintainers cannot review less then 0 changes (for not their
 > co-worker changes). That just math, nothing more.

I think this is not really fair to say. Looking at the Ohloh statistics for Gerrit [3] it can be seen that you got several hundreds of changes in during the last year. This means to say that your changes don't get any attention is simply not correct.

[3] https://classifieds.ohloh.net/p/gerrit/contributors/summary
 
 > But keep blaming
 > me (and other contributors) for maintainers not doing their work,
 > I don't care any more.

That would be very sad. 
As everyone can see from the Ohloh statistics you are by far the most active and productive contributor and your contributions are always very welcome, at least to me! What is not in these statistics is that you are also a very active and valuable reviewer. I want to use this opportunity to say thanks to you for all your efforts!

Patch-set: 11