Update patch set 7
Patch Set 7:

> > Thanks for the report.
 > >
 > > Do you think it would work to concatenate the two Codec files and
 > > store that in the output jar?
 > 
 > I think so, as other people did it successfully it seems; e.g. [1]
 > => [2]. Not sure how to do it myself though yet within this buck
 > context. For sure, having a merged codec service file should work.
 > I can test it myself manually in the meantime; I already tested
 > that a present Lucene50Codec class works.

I don't think it's safe to assume this means it just works. It will probably work for a new installation, since that uses the 5.0 on-disk format, so it only needs the newest codec. For upgrading from an older secondary index, we need to be able to read older index versions. That's the whole purpose of including the backward codecs in the Gerrit build in the first place.

 > Merge-wise, you might
 > know what would be the best approach here.
 
I don't mind writing the buck glue to merge the two files. But I would like reproduction instructions because it'll make it easier to verify :)

 > [1] http://stackoverflow.com/questions/29614796/how-do-i-correctly-output-meta-inf-when-libraries-use-spis-using-gradle
 > =>
 > [2] http://cmoz.me/blog/2014/11/service-files-uber-jars-and-gradle/

Patch-set: 7