Update patch set 3
Patch Set 3:

> > With this change we no longer see the stack of related changes
 > when
 > > looking at the change at the bottom of the stack.
 > 
 > For example look at [1] and see that it is dependent on [2].  But
 > if you click on [2] the tab disappears and you no longer see [1] in
 > the list.
 > 
 > [1] https://gerrit-review.googlesource.com/#/c/69814/2
 > [2] https://gerrit-review.googlesource.com/#/c/69830/1

That is kind of expected. Because when looking at [2] and submit it, it was submitted alone without [1] as [1] sits on top of [2] not coupled to [1] via a topic. So you would submit [2] alone. Would it make more sense to have one tab which shows only [2] as its only entry?

If you want to know that [1] sits on top of [2] you better inspect the related changes tab as that will show the surrounding landscape of the change, while the submitted together tab will show exactly the changes which would be submitted if you press submit at that change.

And having a tab with only one entry is bad, so why show it at all?

Patch-set: 3
Reviewer: Gerrit User 1017837 <1017837@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0