Update patch set 1
Patch Set 1: Code-Review-1 Verified-1

This change is a good example, why LC label should be dropped on gerrit-review. While LC+1 is not voted for months on (my) sane changes, LC+1 is set on this change where LC-1 should be used.

Or have I missed that all Gerrit stakeholders are ready to upgrade to Java 8? As Jetty 9.3.1 depends on Java8, according to this Jesse McConnell's announcement:

  Yes, this is the intention for Jetty 9.3.x moving forward.  We had 
  been operating under the assumption we would support Java 7 but after 
  looking at what is going on in the SSL areas of the jdk with new 
  features like SNI, HTTP/2 becoming a standard and requiring Java 8, 
  and things like we decided to pull that trigger now.

Anyway, all tests are failing with java version mismatch:

  java.lang.UnsupportedClassVersionError: org/eclipse/jetty/server/RequestLog : 
  Unsupported major.minor version 52.0

Patch-set: 1
Label: Code-Review=-1
Label: Verified=-1