Update patch set 1
Patch Set 1:

> > This change is a good example, why LC label should be dropped on
 > > gerrit-review.
 > 
 > I think you're misunderstanding what the Library-Compliance label
 > is for.

As the matter of fact, LC was introduced as Edwin worked on xdoc feature in gerrit core and tried to introduce number of dependencies that weren't acceptable for Google. That was the birth day of LC on gerrit review. As the matter of fact, where are dozens and hundreds of changes that were rejected with LC-1, because someone uploaded (or upgraded) some libraries that are not acceptabe by Google? I'm aware only about this one: [1].

Also from a contributor point of view, why do you think this is attractive and sexy, to do all the work, polish a change, test it, upload it, address review comment and when all these is done ask Google team in change comment or open issues in issue tracker and ask again and again to please vote on LC???

* https://gerrit-review.googlesource.com/67793

Patch-set: 1