Update patch set 1
Patch Set 1:

I was thinking about that, too.  Question: is there already a test to ensure ETags are properly rotated when ChangeResource updates? I'm thinking about this, and the problem I'm running into would be determining what to change to ensure ETags are changed as a result.

Patch-set: 1
Label: Verified=0