Update patch set 7
Patch Set 7:

> Another way would be to pass an additional optional list to
 > MergeSuperSet which will populate the list with changes for you if
 > it's not null, no need to put mutable stuff into a ChangeSet then.

In a public API I'd rather not. The idiomatic Java way to do multiple return values is to return an object of a type that has multiple fields/methods for the pieces you want. Right now that type is called ChangeSet.

Of course, ChangeSet as it is today is not really like what it was originally created for, which was back when it needed to be hashable to handle the multilock stuff. Maybe it shouldn't be an @AutoValue anymore.

Patch-set: 7