Update patch set 9
Patch Set 9:

Note: I split this series into two topics.  The changes in this topic "events" work as expected, as far as I can see with basic testing of stream-events and the replication plugin.

I moved the remaining changes to the "events2" topic.  They will need rebase and conflict resolution, which I think can be done separately from this initial subset of changes in the "events" topic.

Patch-set: 9