Update patch set 28
Patch Set 28:

> Works as expected but there's one UI glitch that I'm not sure if
 > it's a separate issue.
 > 
 > When editing a new empty document, the margin line is only
 > displayed on the first line [1].  As more lines are added to the
 > document, the margin is extended [2].
 > 
 > Compare for example to gedit where the margin is shown on the
 > entire view [3].
 > 
 > 
 > [1] http://imgur.com/zbR1zq8,iXaVkj6,oB5iXel#0
 > [2] http://imgur.com/zbR1zq8,iXaVkj6,oB5iXel#1
 > [3] http://imgur.com/zbR1zq8,iXaVkj6,oB5iXel#2

I see. It would be interesting to invetigate margin behaviour in Codemirror (read only) when only 1 line exists in the file. Is the margin is only visible for the existing line in thi case too? Then I would say this is unrelated issues to inline edit, and can be probably addressed later.

Patch-set: 28