Update patch set 3
Patch Set 3: Code-Review-1

My only suggestion would be to use the replication queue to create the initial project only -- that logic should exist anyway; it'd just be a matter of triggering a replication to each remote.

Was discussing this with David offline, and it may be a problem if there's nothing to replicate initially, but this approach could lead to spawning an indefinite number of threads.  The replication queue should honor the same limits for creation, I think...

Patch-set: 3
Reviewer: Gerrit User 1012612 <1012612@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=-1