Update patch set 2
Patch Set 2:

@ALL: this is a quite serious problem when managing plugins remotely :-( 
I would need to make a demo of a simplified plugin management console ... but this problem is a show-stopper for my talk next week.

Do you have time to review it in the next few days?

In the meantime I will pull this change manually in gerrit-ci.gerritforge.com

Thank you in advance :-)

Patch-set: 2