Update patch set 1
Patch Set 1:

I looked at the SuggestReviewer class and that has no tests. So following that as a bad example this also doesn't have tests (yet). How would I add tests to this with both a good coverage (testing what we want to test) without being affected by the other tests too much? (lots of other tests add topics, which we want to ignore for this test).

Patch-set: 1
Label: Verified=0