Update patch set 4
Patch Set 4: Code-Review+1

Luca, 
The online reindexer was throttled by the Gerrit core to avoid the overload since it was introduced in Gerrit 2.9. On the first implementation the online reindexer was using only one thread. With Gerrit 2.10 this was changed to be throttled using the number of threads defined by changeMerge.threadPoolSize, and finally the last version of the online reindexer with Gerrit 2.11 changed this throttling to be done using index.batchThreads.

In our case with a large instance of almost 700k changes upon upgrading to 2.11 with 6 threads, the online reindexer took almost 7-8 hours to reindex this amount of changes using 6 threads. The load on the Gerrit instance during these 7-8 hours was not significant since we were throttling this to only 6 threads.

Thus there is no overload generated from the online reindexer if the instance is well tuned and the index.batchThreads is appropriately configured based on the number of CPUs on your master instance.

Now for the proposed extension to braodcast an event upon reindexing a change. Since the information broadcasted is already in memory, then there is no CPU expensive operation. Compared to calculating the mergeability check of a change which is performed anyways by the online reindexer, broadcasting the change number which is reindexed is negligable. 

You know there is at least 3 events broadcasted for replicating changes to a single destination. In our instance we have 7 slaves and neither CPU nor Memory was affected by 21 events for replicating single change.

We need to separate here the concerns, i.e. for me I think there is no side effect on Gerrit core with this extension point added, and thus I think we can proceed with merging this extension point. If you still have concerns regarding the plugin and the efficiency of the sync-index plugin we can move these concerns to the plugin once submitted to open source. 

By the way we have this plugin in production for the past 3 months, and there was no CPU/Memory load caused by the plugin in a fairly busy instance.  This extension point is fixing only a bug we found with a corner case, we plan to open source the plugin with this bug fix once this extension is in place.

Patch-set: 4
Reviewer: Gerrit User 1012702 <1012702@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=+1