Update patch set 1
Patch Set 1:

> one can argue that the problem is actually a feature: it prevents
 > people from pushing pointless patchsets.

This feature is still enabled. If the commit that only modified the timestamp does not have a child commit chained, no patch set will be created.

 > n.b., see my comment on the issue if you didn't notice it yet.

I saw your comment and David was trying to fix it as well [1] and it seems like a documentation update is also needed?

[1] https://gerrit-review.googlesource.com/#/c/47930/

Patch-set: 1