Update patch set 1
Patch Set 1:

> NoteDB - Gerrit-CI Build: https://gerrit-ci.gerritforge.com/job/Gerrit-verifier-notedb/1285/consoleText

I haven't fully nailed this down but I think it's because BatchUpdate calls ctx.getChangeUpdate().commit() on l. 525, and that ChangeUpdate can't be reused in this way because it was already written at SubmitStrategyOp.java on l. 209.

Basically BatchUpdate has to do something different if we decide to use BatchMetaDataUpdate for the ChangeUpdate. But how to inform it we intend to do this?

And the other not-really-gonna-do-it option is just finally get around to rewriting the whole MetaDataUpdate thing to not be so confusing.

Patch-set: 1
Label: Verified=0