Update patch set 10
Patch Set 10:

I agree passing the already populated ChangeData would have lots of benefits. The problem with this approach is we would need to bring many other dependencies [1] to gerrit-extension-api to the extension to be able to pass this object:

[1] '//gerrit-common:client',
    '//gerrit-common:server',
    '//gerrit-reviewdb:client',
    '//gerrit-reviewdb:server',
    '//gerrit-server:server'

and my question is, again, do we really want to do this?

Patch-set: 10
Label: Verified=0