Update patch set 4
Patch Set 4:

(1 comment)

@DavidP
I've add a few more tests to answer your questions.  For system groups Gerrit will not allow the creation of a group with the same name (case insensitive verification).  However for internal/custom Groups it is only a case sensitive verification so users can create internal groups that have the same name with different cases.  It's probably debatable whether that's a bug but as it stands the rest and ssh APIs are consistent in detecting duplicate groups.  If you feel that this is a bug it's probably better to fix that in a separate change.

Patch-set: 4
Label: Verified=0