Update patch set 11
Patch Set 11:

> I agree passing the already populated ChangeData would have lots of
 > benefits. The problem with this approach is we would need to bring
 > many other dependencies [1] to gerrit-extension-api to the
 > extension to be able to pass this object:
 > 
 > [1] '//gerrit-common:client',
 > '//gerrit-common:server',
 > '//gerrit-reviewdb:client',
 > '//gerrit-reviewdb:server',
 > '//gerrit-server:server'
 > 
 > and my question is, again, do we really want to do this?

 > I agree passing the already populated ChangeData would have lots of
 > benefits. The problem with this approach is we would need to bring
 > many other dependencies [1] to gerrit-extension-api to the
 > extension to be able to pass this object:
 > 
 > [1] '//gerrit-common:client',
 > '//gerrit-common:server',
 > '//gerrit-reviewdb:client',
 > '//gerrit-reviewdb:server',
 > '//gerrit-server:server'
 > 
 > and my question is, again, do we really want to do this?

Yeah, my suggestion was probably premature, I agree that we can't expose all those things in the extension API. What we need is something like ChangeData with lazy loading semantics but a well-defined API surface area. If someone has some ideas about how to implement that in a sane, performant, supportable way, go for it, but I don't have the bandwidth to deal with that now.

Patch-set: 11