Update patch set 4
Patch Set 4:

(2 comments)

> but we normaly only upload fixes and not new features to stable branches.

Well, it's more of a refactoring to enable fixing a bug, than a new feature.

Having said that, it's not a trivial refactor. We should check if this is going to merge cleanly up to master.  If not, I think I'd prefer to just do it on master instead of stable-2.12.

Patch-set: 4