Update patch set 1
Patch Set 1:

This was uglier than I might have liked but what can you do. We're stuck for now with the mix of manual/automatic/assisted injection that we currently have in notedb.

A few tests are failing (but not ChangeNotesTest :), will fix up if the general implementation looks ok.

Patch-set: 1