Update patch set 1
Patch Set 1:

> Looks like this is an alternative approach to the problem being
 > solved in the verification channel changes [1].  I assume you've
 > been talking about that at the hackathon.
 > 
 > [1] https://gerrit-review.googlesource.com/#/q/project:gerrit+branch:master+topic:verification-notification-channel

Yes. The conclusion around the table was to add minimal support for the change comment filtering in Gerrit core. This change is part of it. UI changes is not part of this change. We are going to extend the database schema and add this flag to the Notedb. It was discussed an option to only enable this feature for Notedb, like hashtags, but it was rejected.

Patch-set: 1