Update patch set 1
Patch Set 1:

The thing why I did not add the functionality on the server side is that I'd have to change the exposed API to accept and return multiple keys. If this is acceptable for you I'd be glad to implement it on the server is the implementation would indeed be much simpler.

Patch-set: 1
Reviewer: Gerrit User 1024958 <1024958@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0