Update patch set 2
Patch Set 2:

I have to rework this significantly after the test failure. The issue is that InitIndex needs to know all the latest schema versions, but the injector using to create it is long before the IndexModule is (or can be) installed. So we need to split IndexDefinition into the static Schema<?> definitions (basically what ChangeSchemas was prior to this change) and the dynamic stuff that can only be created under the IndexModule.

Patch-set: 2
Label: Verified=0