Update patch set 1
Patch Set 1:

> > When you say "predicates" I think you don't mean Prolog
 > predicates
 > > but just these Java implementations of "MaxWithBlock", etc..
 > exposed
 > > as reusable components for plugins?
 > 
 > No, I actually mean Prolog predicates. Currently you can write
 > Prolog submit rules containing predicates that correspond to the
 > existing function names:
 > 
 > https://gerrit-review.googlesource.com/Documentation/prolog-cookbook.html#_default_submit_rule_implementation
 > 
 > If you move MaxWithBlock to be implemented entirely in Java, you
 > need to make sure existing Prolog rules don't break.

Discussed in person. Plugins that want to hook into the existing submit rule functions can call the public classes exposed from core. We think it's safest to not touch the existing Prolog implementations (but we could if necessary).

The only risk here is that the implementation/definition of one of the existing functions changes, and we might forget to make the corresponding change to the Prolog implementation of that function. But we don't think that is likely to happen. 

If new functions are added to core, they will not show up automatically in Prolog. That's probably fine, it's not hard to wire them up if/when that happens.

We could get very clever and automatically generate predicate implementations for each of the functions in core, but the cost of implementing that does not outweigh the benefit of pushing this series through.

Patch-set: 1