Update patch set 2
Patch Set 2:

Unfortunately this ignores the fact that some internal implementations of Gerrit use a different GerritServerConfig which can be interrogated outside of a file system (for example, through a database). So your assumption that changes cannot have an effect at runtime don't hold.

Please remove your -1 so this change can be merged. Not only has it been tested, it is done in the same way as the existing SSH implementation it is based on a falsifiable assumption.

Patch-set: 2