Update patch set 1
Patch Set 1:

> > > So with this revert you can push related changes to a project
 > > with
 > > > createNewChangeForAllNotInTarget = true but the actual
 > > > functionality of createNewChangeForAllNotInTarget = true is
 > > broken
 > > > again so it makes no sense having that configured?
 > >
 > > Pushing dependent changes is very much core Gerrit functionality,
 > > so if we're going to have one of those two things broken, it
 > > shouldn't be that.
 > 
 > Yes exactly. I also think that not being able to push 2 changes
 > which depend on each other is worse than the other issue, hence the
 > revert.

I agree that it's a worse issue. But isn't there a catch 22 here? If you choose the setting to create new changes for all not in target you are now able to push a chain of changes. But you're not able to "create a new change for all not in target", am I right?
So if you select this setting you can now create a dependency chain of changes, but the setting has no effect.
So the result of this Revert Reapply Revert... is pretty much the same as if you set createNewChangeForAllNotInTarget = false.

My question is: Is it better to remove functionality than to keep it with a known bug?

Patch-set: 1