Update patch set 1
Patch Set 1:

Thanks for the discussion!

I did not mean to imply a shift in responsibility to core developers,
but rather as keep the responsibility as is.

In case someone is setting up a CI it *may* be useful to know about all the plugins that can easily be built with BUCK as it doesn't add more programs to invoke for CI. For that we could either include only BUCK plugins, or my preferred way would be to add another label e.g.  "build-BUCK", such that you could obtain them with

    git clone --init-submodules=":(attr:build-BUCK)" <Gerrits Url>

or
    
    git submodule init ":(attr:build-BUCK)"
    git submodule update

for existing repositories.

My main motivation for this change is to have a real world example of a repository with submodules, but a lot of diverse interests for these submodules (some people don't want any, some want the "defaults", others want "all", others want the intersection of "build-BUCK" and "lang-java").

Having Gerrit as this superrepo with the collection of submodules makes a lot of sense, too, as we can work with the superproject subscription feature, which is going to be enhanced soon, too.

So my main motivation is a bit off from what I wrote as the commit message as the commit message caters to the project needs. Why should Gerrit as a community be interested in this change? And for that I think a better organisation of plugins is a good selling point.
By organisation I mean both

* having the possibility to keep all data in one git repository, and have it organized over time, i.e. we can view when a plugin upgraded or fixed itself to be compatible with Gerrit again

* having an overview of existing plugins via labels in the .gitattributes file. This is comparable to hashtags (one submodule can have multiple tags, you can search/select by tag). This allows for a better birds eye perspective on the community around the core.

As this is dogfooding some new features of Git, this may also be a bit of a burden as 

* it adds complexity. (The user has to decide what they want. It's not a binary "I want [no] submodules" via

    git clone --[no-]recurse-submodules

any more, but a finer selection means a little more overhead. We would need to document that most people would be interested in the "default" submodules.

* it may lead to confusion for existing members of the project as the workflow changes. ("Why are there so many submodules?")

Overall I think (from the gerrits POV) the benefits outweigh the disadvantages.

Patch-set: 1
Reviewer: Gerrit User 1017837 <1017837@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=0