Update patch set 1
Patch Set 1:

I wonder if this is a good idea, when we are actually trying to move the Prolog submit rules into a plugin [1], which means dropping the dependency to Prolog in Gerrit core.

So for now exporting the Prolog compiler will work, but somewhen in the future when Prolog is removed, plugins depending on this will break.

[1] https://gerrit-review.googlesource.com/#/q/project:gerrit+branch:master+topic:non-prolog-submit-rules-gh16

Patch-set: 1