Update patch set 2
Patch Set 2:

> > Hashtag added: blocking-2.13
 > 
 > Is this really a blocker for 2.13?

Without this change (This change is a WIP to highlight the issue), when you update to 2.13, you end up with no change in the index. To work around the issue, you need to manually edit gerrit_index.config, set the latest index ready state to false, restart the server to search from previous index until online reindexer is done.

I think the work around requires too much in-depth gerrit knowledge so we consider releasing 2.13 with this bug.

Patch-set: 2